Add IoHelper::logArchiverDirectoryPath.

+++ src/libsyncengine/requests/serverrequests.cpp
@@ -998,13 +998,22 @@ ExitCode ServerRequests::sendLogToSupport(bool includeArchivedLog, std::function
 
     safeProgressCallback(LogUploadState::Archiving, 0);
 
-    if (bool found = false; !ParmsDb::instance()->updateAppState(AppStateKey::LastLogUploadArchivePath, std::string(""), found) || !found) {
+    if (bool found = false;
+        !ParmsDb::instance()->updateAppState(AppStateKey::LastLogUploadArchivePath, std::string(""), found) || !found) {
         LOG_WARN(Log::instance()->getLogger(), "Error in ParmsDb::updateAppState");
         // Do not return here because it is not a critical error, especially in this context where we are trying to send logs
     }
 
-    const SyncPath logUploadTempFolder = CommonUtility::getAppSupportDir() / "logUploadTemp";
+    SyncPath logUploadTempFolder;
     IoError ioError = IoErrorSuccess;
+
+    IoHelper::logArchiverDirectoryPath(logUploadTempFolder, ioError);
+    if (ioError != IoErrorSuccess) {
+        LOG_WARN(Log::instance()->getLogger(),
+                            "Error in IoHelper::logArchiverDirectoryPath : " << Utility::formatIoError(logUploadTempFolder, ioError).c_str());
+        return ExitCodeSystemError;
+    }
+
     IoHelper::createDirectory(logUploadTempFolder, ioError);
     if (ioError == IoErrorDirectoryExists) {  // If the directory already exists, we delete it and recreate it
         IoHelper::deleteDirectory(logUploadTempFolder, ioError);
@@ -1065,9 +1074,9 @@ ExitCode ServerRequests::sendLogToSupport(bool includeArchivedLog, std::function
     long long timestamp =
         std::chrono::time_point_cast<std::chrono::seconds>(std::chrono::system_clock::now()).time_since_epoch().count();
     std::string uploadDate = std::to_string(timestamp);
-    if (bool found = false; !ParmsDb::instance()->updateAppState(AppStateKey::LastSuccessfulLogUploadDate, uploadDate, found) ||
-                            !found || !ParmsDb::instance()->updateAppState(AppStateKey::LastLogUploadArchivePath, std::string(""), found) ||
-                            !found) {
+    if (bool found = false;
+        !ParmsDb::instance()->updateAppState(AppStateKey::LastSuccessfulLogUploadDate, uploadDate, found) || !found ||
+        !ParmsDb::instance()->updateAppState(AppStateKey::LastLogUploadArchivePath, std::string(""), found) || !found) {
         LOG_WARN(Log::instance()->getLogger(), "Error in ParmsDb::updateAppState");
     }
     return ExitCodeOk;
@@ -1093,7 +1102,8 @@ ExitCode ServerRequests::cancelLogToSupport(ExitCause &exitCause) {
         return ExitCodeInvalidOperation;  // The operation is not in progress
     }
 
-    if (bool found = false; !ParmsDb::instance()->updateAppState(AppStateKey::LogUploadState, LogUploadState::CancelRequested, found) || !found) {
+    if (bool found = false;
+        !ParmsDb::instance()->updateAppState(AppStateKey::LogUploadState, LogUploadState::CancelRequested, found) || !found) {
         LOG_WARN(Log::instance()->getLogger(), "Error in ParmsDb::updateAppState");
         return ExitCodeDbError;
     }
